+++ clients/ctl/admin-ui/src/features/dataset/DatasetFieldsTable.tsx
@@ -1,6 +1,7 @@
-import { Box, Table, Tbody, Td, Th, Thead, Tr } from "@fidesui/react";
+import { Box, Table, Tbody, Td, Th, Thead, Tooltip, Tr } from "@fidesui/react";
 import { useDispatch, useSelector } from "react-redux";
 
+import { selectClassificationFieldMap } from "~/features/common/plus.slice";
 import { DatasetField } from "~/types/api";
 
 import IdentifiabilityTag from "../taxonomy/IdentifiabilityTag";
@@ -20,12 +21,15 @@ import { ColumnMetadata, EditableType } from "./types";
 const Cell = ({
   attribute,
   field,
+  classificationField,
 }: {
   attribute: keyof DatasetField;
   field: DatasetField;
+  classificationField?: DatasetField;
 }): JSX.Element => {
   if (attribute === "data_qualifier") {
-    const dataQualifierName = field[attribute];
+    const dataQualifierName = field.data_qualifier;
+
     return (
       <>
         {dataQualifierName ? (
@@ -36,14 +40,30 @@ const Cell = ({
   }
 
   if (attribute === "data_categories") {
+    const classifiedCategories = classificationField?.data_categories ?? [];
+    const assignedCategories = field.data_categories ?? [];
+    // Only show the classified categories if none have been directly assigned to the dataset.
+    const categories =
+      assignedCategories.length > 0 ? assignedCategories : classifiedCategories;
+
     return (
-      <>
-        {(field[attribute] ?? []).map((dc) => (
-          <Box key={`${field.name}-${dc}`} mr={2} mb={2}>
-            <TaxonomyEntityTag name={dc} />
-          </Box>
-        ))}
-      </>
+      <Tooltip
+        placement="right"
+        label={
+          // TODO: Related to #724, the design wants this to be clickable but our tooltip doesn't support that.
+          categories === classifiedCategories
+            ? "Fides has generated these data categories for you. You can override them by modifying the field."
+            : ""
+        }
+      >
+        <Box display="inline-block">
+          {categories.map((dc) => (
+            <Box key={`${field.name}-${dc}`} mr={2} mb={2}>
+              <TaxonomyEntityTag name={dc} />
+            </Box>
+          ))}
+        </Box>
+      </Tooltip>
     );
   }
 
@@ -60,6 +80,7 @@ const DatasetFieldsTable = ({ columns }: Props) => {
   const activeFields = useSelector(selectActiveFields);
   const activeField = useSelector(selectActiveField);
   const activeEditor = useSelector(selectActiveEditor);
+  const classificationFieldMap = useSelector(selectClassificationFieldMap);
 
   const handleClose = () => {
     dispatch(setActiveFieldIndex(undefined));
@@ -100,7 +121,11 @@ const DatasetFieldsTable = ({ columns }: Props) => {
             >
               {columns.map((c) => (
                 <Td key={`${c.name}-${field.name}`} pl={0}>
-                  <Cell field={field} attribute={c.attribute} />
+                  <Cell
+                    field={field}
+                    classificationField={classificationFieldMap.get(field.name)}
+                    attribute={c.attribute}
+                  />
                 </Td>
               ))}
             </Tr>